+++ Control/Joint/Effects.hs
@@ -1,6 +1,7 @@
 module Control.Joint.Effects (module Exports) where
 
-import Control.Joint.Effects.State as Exports
 import Control.Joint.Effects.Maybe as Exports
 import Control.Joint.Effects.Either as Exports
+import Control.Joint.Effects.State as Exports
+import Control.Joint.Effects.Writer as Exports
 import Control.Joint.Effects.Reader as Exports