+++ Control/Joint/Effects/Reader.hs
@@ -20,16 +20,6 @@ instance Applicative (Reader e) where
 instance Monad (Reader e) where
 	Reader g >>= f = Reader $ \e -> run (f (g e)) e
 
-instance Functor u => Functor (TU ((->) e) u) where
-	fmap f (TU x) = TU $ \r -> f <$> x r
-
-instance Applicative u => Applicative (TU ((->) e) u) where
-	pure = TU . pure . pure
-	TU f <*> TU x = TU $ \r -> f r <*> x r
-
-instance (Applicative u, Monad u) => Monad (TU ((->) e) u) where
-	TU x >>= f = TU $ \e -> x e >>= ($ e) . run . f
-
 instance Composition (Reader e) where
 	type Primary (Reader e) a = (->) e a
 	run (Reader x) = x
@@ -39,6 +29,16 @@ instance Transformer (Reader e) where
 	embed x = TU . const $ x
 	build x = TU $ pure <$> run x
 	unite = TU
+
+instance Functor u => Functor (TU ((->) e) u) where
+	fmap f (TU x) = TU $ \r -> f <$> x r
+
+instance Applicative u => Applicative (TU ((->) e) u) where
+	pure = TU . pure . pure
+	TU f <*> TU x = TU $ \r -> f r <*> x r
+
+instance (Applicative u, Monad u) => Monad (TU ((->) e) u) where
+	TU x >>= f = TU $ \e -> x e >>= ($ e) . run . f
 
 instance Modulator (Reader e) where
 	f -<$>- (TU x) = TU $ f <$> x