+++ Control/Joint/Abilities/Transformer.hs
@@ -6,16 +6,12 @@ import Control.Joint.Abilities.Composition (Composition (Primary, run))
 class Composition t => Transformer t where
 	{-# MINIMAL embed, build, unite #-}
 	type Schema (t :: * -> *) (u :: * -> *) = (r :: * -> *) | r -> t u
-	embed :: Functor u => u ~> Schema t u
-	build :: Applicative u => t ~> Schema t u
-	unite :: Primary (Schema t u) a -> Schema t u a
+	embed :: Functor u => u ~> t :> u
+	build :: Applicative u => t ~> t :> u
+	unite :: Primary (Schema t u) a -> (t :> u) a
 
 infixr 0 :>
-newtype (:>) t u a = T { trans :: Transformer t => (Schema t u a) }
-
-instance (Composition (Schema t u), Transformer t) => Composition (t :> u) where
-	type Primary (t :> u) a = Primary (Schema t u) a
-	run (T x) = run x
+newtype (:>) t u a = T { trans :: Transformer t => Schema t u a }
 
 instance Functor (Schema t u) => Functor (t :> u) where
 	fmap f (T x) = T $ f <$> x
@@ -27,8 +23,6 @@ instance (Transformer t, Applicative (Schema t u)) => Applicative (t :> u) where
 instance (Transformer t, Monad (Schema t u)) => Monad (t :> u) where
 	T x >>= f = T $ x >>= trans . f
 
-instance (Composition (Schema t u), Transformer t) => Transformer ((:>) t u) where
-    type Schema (t :> u) v = t :> u :> v
-    embed = embed
-    build = build
-    unite = unite
+instance (Composition (Schema t u), Transformer t) => Composition (t :> u) where
+	type Primary (t :> u) a = Primary (Schema t u) a
+	run (T x) = run x