+++ Control/Joint/Abilities/Modulator.hs
@@ -1,4 +1,0 @@
-module Control.Joint.Abilities.Modulator where
-
-import Control.Joint.Core (type (~>))
-import Control.Joint.Abilities.Transformer ((:>))